Split server from frontend route
